Add files via upload
We now insert functionalities into our program so that it will now add the lecturer's information from name to hours worked. The insert of that information will now appear in the program coordinators page when he/she accesses it, and it will be their decision if they want to approve or decline your claim depending if the hours worked are indeed correct.